+++ lib/legion/extensions/dynatrace/hub.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+require_relative 'hub/runners/hub'
+require_relative 'helpers/client'
+require_relative 'base_client'
+
+module Legion
+  module Extensions
+    module Dynatrace
+      module Hub
+        class Client < BaseClient
+          include Runners::Hub
+        end
+      end
+    end
+  end
+end